Add project: Peter + solution
Add "Peter" project, tried some of the things we learned from the course. Feel free to look around, edit or anyhow improve the code. Would be neat! I have some issues addressed in the controller that could be taken care of if anyone has some sparetime!
BTW - These files might be pushed into the wrong directive, is this directly into the mvc201701 folder? It needs to be placed as a project parallel to mvc201701.